Restructured pdfviewer(wx).py
